requirements.txt

django-allauth==0.35.0
django-appconf==1.0.3
django-bootstrap3==11.0.0
django-bootstrap4==0.0.6
django-braces==1.13.0
django-crispy-forms==1.7.2
django-datetime-widget==0.9.3
django-environ==0.4.4
django-model-utils==3.1.1
django-select2==6.0.2
django-simple-history==2.3.0
django-tenants==2.0.0
djangorestframework==3.8.2
docutils==0.14
idna==2.8
imagesize==1.1.0
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
MarkupSafe==1.1.1
oauthlib==3.0.1
packaging==19.0
parso==0.3.4
pickleshare==0.7.5
Pillow==5.1.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
pycosat==0.6.3
pycparser==2.19
Pygments==2.3.1
pyOpenSSL==18.0.0
pyparsing==2.3.1
PySocks==1.6.8
python3-openid==3.1.0
pytz==2018.4
pywin32==223
redis==3.2.0
regex==2019.2.21
requests==2.21.0
requests-oauthlib==1.2.0
ruamel-yaml==0.15.46
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-websupport==1.1.0
traitlets==4.3.2
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==3.3.1
win-inet-pton==1.0.1
wincertstore==0.2